Split prm_test_helper.rb
